requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
passlib[bcrypt]
pytest
httpx
psycopg2-binary
python-jose
pytest-playwright
